# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "C2D Admin API",
    "version": "1.0.0",
    "description": "API для админки ИИ-агента и рассылки сообщений"
  },
  "servers": [
    {"url": "/api", "description": "API Server"}
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/clients": {
      "get": {
        "tags": ["Clients"],
        "summary": "Список клиентов",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "search", "in": "query", "schema": {"type": "string"}, "description": "Поиск"}],
        "responses": {"200": {"description": "Список клиентов"}}
      },
      "post": {
        "tags": ["Clients"],
        "summary": "Создать клиента",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phone"],
                "properties": {
                  "phone": {"type": "string", "example": "+79001234567"},
                  "first_name": {"type": "string"},
                  "last_name": {"type": "string"},
                  "email": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {"201": {"description": "Клиент создан"}}
      }
    },
    "/clients/{id}": {
      "get": {
        "tags": ["Clients"],
        "summary": "Получить клиента по ID",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Данные клиента"}}
      },
      "put": {
        "tags": ["Clients"],
        "summary": "Обновить клиента",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Клиент обновлен"}}
      },
      "delete": {
        "tags": ["Clients"],
        "summary": "Удалить клиента",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Клиент удален"}}
      }
    },
    "/clients/import": {
      "post": {
        "tags": ["Clients"],
        "summary": "Импорт клиентов из JSON",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["clients"],
                "properties": {
                  "clients": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "phone": {"type": "string", "example": "+79001234567"},
                        "first_name": {"type": "string"},
                        "last_name": {"type": "string"}
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {"200": {"description": "Импорт завершен"}}
      }
    },
    "/clients/export": {
      "get": {
        "tags": ["Clients"],
        "summary": "Экспорт клиентов",
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "JSON со списком клиентов"}}
      }
    },
    "/dialogs": {
      "get": {
        "tags": ["Dialogs"],
        "summary": "Список диалогов",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "status", "in": "query", "schema": {"type": "string"}},
          {"name": "campaign_id", "in": "query", "schema": {"type": "integer"}}
        ],
        "responses": {"200": {"description": "Список диалогов"}}
      }
    },
    "/dialogs/{id}": {
      "get": {
        "tags": ["Dialogs"],
        "summary": "Получить диалог по ID",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Данные диалога"}}
      }
    },
    "/dialogs/{id}/status": {
      "post": {
        "tags": ["Dialogs"],
        "summary": "Обновить статус диалога",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {"type": "string", "example": "ready_to_buy"},
                  "metadata": {"type": "object"}
                }
              }
            }
          }
        },
        "responses": {"200": {"description": "Статус обновлен"}}
      }
    },
    "/statistics/campaigns": {
      "get": {
        "tags": ["Statistics"],
        "summary": "Статистика по кампаниям",
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Статистика кампаний"}}
      }
    },
    "/statistics/dialogs": {
      "get": {
        "tags": ["Statistics"],
        "summary": "Статистика по диалогам",
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Статистика диалогов"}}
      }
    },
    "/statistics/export": {
      "get": {
        "tags": ["Statistics"],
        "summary": "Экспорт статистики",
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "JSON с полной статистикой"}}
      }
    },
    "/webhooks/ai-agent": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Webhook от ИИ-агента",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["chat_id", "status"],
                "properties": {
                  "chat_id": {"type": "string", "example": "chat_123"},
                  "status": {"type": "string", "example": "presentation"},
                  "client_phone": {"type": "string", "example": "+79001234567"},
                  "collected_data": {"type": "object"}
                }
              }
            }
          }
        },
        "responses": {"200": {"description": "OK"}}
      }
    },
...
  }
}